Break SAU tests out into their own file.
